Resource Access

+++ Services/RestaurantService.cs
@@ -1,4 +1,5 @@
-﻿using AutoMapper;
+﻿using System;
+using AutoMapper;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Logging;
 using RestaurantAPI.Exceptions;
@@ -6,6 +7,9 @@ using RestaurantAPI.Entities;
 using RestaurantAPI.Models;
 using System.Collections.Generic;
 using System.Linq;
+using System.Security.Claims;
+using Microsoft.AspNetCore.Authorization;
+using RestaurantAPI.Authorization;
 
 namespace RestaurantAPI.Services
 {
@@ -13,30 +17,44 @@ namespace RestaurantAPI.Services
     {
         RestaurantDto GetById(int id);
         IEnumerable<RestaurantDto> GetAll();
-        int Create(CreateRestaurantDto dto);
-        void Delete(int id);
-        void Update(int id, UpdateRestaurantDto dto);
+        int Create(CreateRestaurantDto dto, int userId);
+        void Delete(int id, ClaimsPrincipal user);
+        void Update(int id, UpdateRestaurantDto dto, ClaimsPrincipal user);
     }
     public class RestaurantService : IRestaurantService
     {
         private readonly RestaurantDbContext _dbContext;
         private readonly IMapper _mapper;
         private readonly ILogger<RestaurantService> _logger;
-        public RestaurantService(RestaurantDbContext dbContext, IMapper mapper, ILogger<RestaurantService> logger)
+        private readonly IAuthorizationService _authorizationService;
+
+        public RestaurantService(RestaurantDbContext dbContext, IMapper mapper, ILogger<RestaurantService> logger
+        , IAuthorizationService authorizationService)
         {
             _dbContext = dbContext;
             _mapper = mapper;
             _logger = logger;
+            _authorizationService = authorizationService;
         }
 
-        public void Update(int id, UpdateRestaurantDto dto)
+        public void Update(int id, UpdateRestaurantDto dto, ClaimsPrincipal user)
         {
+            
+
             var restaurant = _dbContext
                .Restaurants
                .FirstOrDefault(r => r.Id == id);
 
             if (restaurant is null)
                 throw new NotFoundException("Restaurant not found");
+
+            var authorizationResult = _authorizationService.AuthorizeAsync(user, restaurant,
+                new ResourceOperationsRequirement(ResourceOperation.Update)).Result;
+
+            if (!authorizationResult.Succeeded)
+            {
+                throw new ForbidException();
+            }
 
             restaurant.Name = dto.Name;
             restaurant.Description = dto.Description;
@@ -45,7 +63,7 @@ namespace RestaurantAPI.Services
             _dbContext.SaveChanges();
         }
 
-        public void Delete(int id)
+        public void Delete(int id, ClaimsPrincipal user)
         {
             //jeżeli klient wyśle zapytanie do akcji dla usuwania restauracji
             //zanim się ona wykona zostanie zapisana do pliku z logami
@@ -56,6 +74,13 @@ namespace RestaurantAPI.Services
 
             if (restaurant is null)
                 throw new NotFoundException("Restaurant not found");
+            
+            var authorizationResult = _authorizationService.AuthorizeAsync(user, restaurant,
+                new ResourceOperationsRequirement(ResourceOperation.Delete)).Result;
+            if (!authorizationResult.Succeeded)
+            {
+                throw new ForbidException();
+            }
 
             _dbContext.Restaurants.Remove(restaurant);
             _dbContext.SaveChanges();
@@ -89,9 +114,11 @@ namespace RestaurantAPI.Services
             var restaurantsDtos = _mapper.Map<List<RestaurantDto>>(restaurant);
             return restaurantsDtos;
         }
-        public int Create(CreateRestaurantDto dto)
+        public int Create(CreateRestaurantDto dto,int userId)
         {
+
             var restaurant = _mapper.Map<Restaurant>(dto);
+            restaurant.CreatedById = userId;
             _dbContext.Restaurants.Add(restaurant);
             _dbContext.SaveChanges();
             return restaurant.Id;
